Funções para definição dos valores ideais
Implementação das funções de definição dos valores ideais para fósforo, potássio, cálcio, enxofre e magnésio

+++ Solos-Arquitetura/src/main/java/com/solos/arquitetura/calcio_magnesio/CorrecaoCalcioMagnesio.java
@@ -10,41 +10,41 @@ package com.solos.arquitetura.calcio_magnesio;
  * @author Rique
  */
 public class CorrecaoCalcioMagnesio {
-    private float calcio_desejado_na_CTC;
-    private float fonte_corretivo_a_usar;
-    private float PRTV;
-    private float teor_CaO;
+    private double calcio_desejado_na_CTC;
+    private double fonte_corretivo_a_usar;
+    private double PRTV;
+    private double teor_CaO;
 
     // <editor-fold defaultstate="collapsed" desc=" Getters & Setters ">
-    public float getCalcio_desejado_na_CTC() {
+    public double getCalcio_desejado_na_CTC() {
         return calcio_desejado_na_CTC;
     }
 
-    public void setCalcio_desejado_na_CTC(float calcio_desejado_na_CTC) {
+    public void setCalcio_desejado_na_CTC(double calcio_desejado_na_CTC) {
         this.calcio_desejado_na_CTC = calcio_desejado_na_CTC;
     }
 
-    public float getFonte_corretivo_a_usar() {
+    public double getFonte_corretivo_a_usar() {
         return fonte_corretivo_a_usar;
     }
 
-    public void setFonte_corretivo_a_usar(float fonte_corretivo_a_usar) {
+    public void setFonte_corretivo_a_usar(double fonte_corretivo_a_usar) {
         this.fonte_corretivo_a_usar = fonte_corretivo_a_usar;
     }
 
-    public float getPRTV() {
+    public double getPRTV() {
         return PRTV;
     }
 
-    public void setPRTV(float PRTV) {
+    public void setPRTV(double PRTV) {
         this.PRTV = PRTV;
     }
 
-    public float getTeor_CaO() {
+    public double getTeor_CaO() {
         return teor_CaO;
     }
 
-    public void setTeor_CaO(float teor_CaO) {
+    public void setTeor_CaO(double teor_CaO) {
         this.teor_CaO = teor_CaO;
     }
     // </editor-fold>
