+++ Solos-Arquitetura/src/main/java/com/solos/arquitetura/fosforo/CorrecaoFosforo.java
@@ -10,33 +10,33 @@ package com.solos.arquitetura.fosforo;
  * @author Rique
  */
 public class CorrecaoFosforo {
-    private float teor_desejado;
-    private float fonte_a_utilizar;
-    private float eficiencia;
+    private double teor_desejado;
+    private double fonte_a_utilizar;
+    private double eficiencia;
     
     // <editor-fold defaultstate="collapsed" desc=" Getters & Setters ">
 
-    public float getTeor_desejado() {
+    public double getTeor_desejado() {
         return teor_desejado;
     }
 
-    public void setTeor_desejado(float teor_desejado) {
+    public void setTeor_desejado(double teor_desejado) {
         this.teor_desejado = teor_desejado;
     }
 
-    public float getFonte_a_utilizar() {
+    public double getFonte_a_utilizar() {
         return fonte_a_utilizar;
     }
 
-    public void setFonte_a_utilizar(float fonte_a_utilizar) {
+    public void setFonte_a_utilizar(double fonte_a_utilizar) {
         this.fonte_a_utilizar = fonte_a_utilizar;
     }
 
-    public float getEficiencia() {
+    public double getEficiencia() {
         return eficiencia;
     }
 
-    public void setEficiencia(float eficiencia) {
+    public void setEficiencia(double eficiencia) {
         this.eficiencia = eficiencia;
     }
     //</editor-fold>
