+++ Solos-Arquitetura/src/main/java/com/solos/arquitetura/fosforo/FontesFosforo.java
@@ -10,114 +10,114 @@ package com.solos.arquitetura.fosforo;
  * @author Rique
  */
 public class FontesFosforo {
-    private float superfosfato_simples;
-    private float superfosfato_triplo;
-    private float MAP;
-    private float DAP;
-    private float yoorin;
-    private float fosfato_arad;
-    private float fosfato_gafsa;
-    private float fosfato_daoui;
-    private float fosfato_patos_minas;
-    private float escoria_de_thomas;
-    private float acido_fosforico;
-    private float multif_magnesiano;
+    private double superfosfato_simples;
+    private double superfosfato_triplo;
+    private double MAP;
+    private double DAP;
+    private double yoorin;
+    private double fosfato_arad;
+    private double fosfato_gafsa;
+    private double fosfato_daoui;
+    private double fosfato_patos_minas;
+    private double escoria_de_thomas;
+    private double acido_fosforico;
+    private double multif_magnesiano;
     
     // <editor-fold defaultstate="collapsed" desc=" Getters & Setters ">
 
-    public float getSuperfosfato_simples() {
+    public double getSuperfosfato_simples() {
         return superfosfato_simples;
     }
 
-    public void setSuperfosfato_simples(float superfosfato_simples) {
+    public void setSuperfosfato_simples(double superfosfato_simples) {
         this.superfosfato_simples = superfosfato_simples;
     }
 
-    public float getSuperfosfato_triplo() {
+    public double getSuperfosfato_triplo() {
         return superfosfato_triplo;
     }
 
-    public void setSuperfosfato_triplo(float superfosfato_triplo) {
+    public void setSuperfosfato_triplo(double superfosfato_triplo) {
         this.superfosfato_triplo = superfosfato_triplo;
     }
 
-    public float getMAP() {
+    public double getMAP() {
         return MAP;
     }
 
-    public void setMAP(float MAP) {
+    public void setMAP(double MAP) {
         this.MAP = MAP;
     }
 
-    public float getDAP() {
+    public double getDAP() {
         return DAP;
     }
 
-    public void setDAP(float DAP) {
+    public void setDAP(double DAP) {
         this.DAP = DAP;
     }
 
-    public float getYoorin() {
+    public double getYoorin() {
         return yoorin;
     }
 
-    public void setYoorin(float yoorin) {
+    public void setYoorin(double yoorin) {
         this.yoorin = yoorin;
     }
 
-    public float getFosfato_arad() {
+    public double getFosfato_arad() {
         return fosfato_arad;
     }
 
-    public void setFosfato_arad(float fosfato_arad) {
+    public void setFosfato_arad(double fosfato_arad) {
         this.fosfato_arad = fosfato_arad;
     }
 
-    public float getFosfato_gafsa() {
+    public double getFosfato_gafsa() {
         return fosfato_gafsa;
     }
 
-    public void setFosfato_gafsa(float fosfato_gafsa) {
+    public void setFosfato_gafsa(double fosfato_gafsa) {
         this.fosfato_gafsa = fosfato_gafsa;
     }
 
-    public float getFosfato_daoui() {
+    public double getFosfato_daoui() {
         return fosfato_daoui;
     }
 
-    public void setFosfato_daoui(float fosfato_daoui) {
+    public void setFosfato_daoui(double fosfato_daoui) {
         this.fosfato_daoui = fosfato_daoui;
     }
 
-    public float getFosfato_patos_minas() {
+    public double getFosfato_patos_minas() {
         return fosfato_patos_minas;
     }
 
-    public void setFosfato_patos_minas(float fosfato_patos_minas) {
+    public void setFosfato_patos_minas(double fosfato_patos_minas) {
         this.fosfato_patos_minas = fosfato_patos_minas;
     }
 
-    public float getEscoria_de_thomas() {
+    public double getEscoria_de_thomas() {
         return escoria_de_thomas;
     }
 
-    public void setEscoria_de_thomas(float escoria_de_thomas) {
+    public void setEscoria_de_thomas(double escoria_de_thomas) {
         this.escoria_de_thomas = escoria_de_thomas;
     }
 
-    public float getAcido_fosforico() {
+    public double getAcido_fosforico() {
         return acido_fosforico;
     }
 
-    public void setAcido_fosforico(float acido_fosforico) {
+    public void setAcido_fosforico(double acido_fosforico) {
         this.acido_fosforico = acido_fosforico;
     }
 
-    public float getMultif_magnesiano() {
+    public double getMultif_magnesiano() {
         return multif_magnesiano;
     }
 
-    public void setMultif_magnesiano(float multif_magnesiano) {
+    public void setMultif_magnesiano(double multif_magnesiano) {
         this.multif_magnesiano = multif_magnesiano;
     }
    // </editor-fold>
