+++ Solos-Arquitetura/src/main/java/com/solos/arquitetura/potassioo/CorrecaoPotassio.java
@@ -10,24 +10,24 @@ package com.solos.arquitetura.potassioo;
  * @author Rique
  */
 public class CorrecaoPotassio {
-    private float participacao_CTC_desejada;
-    private float fonte_a_usar;
+    private double participacao_CTC_desejada;
+    private double fonte_a_usar;
     
     // <editor-fold defaultstate="collapsed" desc=" Getters & Setters ">
 
-    public float getParticipacao_CTC_desejada() {
+    public double getParticipacao_CTC_desejada() {
         return participacao_CTC_desejada;
     }
 
-    public void setParticipacao_CTC_desejada(float participacao_CTC_desejada) {
+    public void setParticipacao_CTC_desejada(double participacao_CTC_desejada) {
         this.participacao_CTC_desejada = participacao_CTC_desejada;
     }
 
-    public float getFonte_a_usar() {
+    public double getFonte_a_usar() {
         return fonte_a_usar;
     }
 
-    public void setFonte_a_usar(float fonte_a_usar) {
+    public void setFonte_a_usar(double fonte_a_usar) {
         this.fonte_a_usar = fonte_a_usar;
     }
     //</editor-fold>
